% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_Qs.R
\name{classify_Qs}
\alias{classify_Qs}
\title{Transfor Qs_rel into binary factor with low and high specific capacity}
\usage{
classify_Qs(x, split_point = 80, class_names = c("low", "high"))
}
\arguments{
\item{x}{vector of Qs_rel values}

\item{split_point}{threshold for classifying numeric Qs_rel values, default: 80}

\item{class_names}{class names, default: c("low", "high")}
}
\description{
Transfor Qs_rel into binary factor with low and high specific capacity
}
