% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pump_test_and_Q_monitoring_data.R
\name{combine_pump_test_and_Q_monitoring_data}
\alias{combine_pump_test_and_Q_monitoring_data}
\title{Combined Pumptest and Q Monitoring Dataset}
\usage{
combine_pump_test_and_Q_monitoring_data(
  df_pump_tests_tidy,
  df_Q_monitoring,
  pump_test_vars = get_pump_test_vars()
)
}
\arguments{
\item{df_pump_tests_tidy}{df_pump_tests_tidy}

\item{df_Q_monitoring}{df_Q_monitoring}

\item{pump_test_vars}{default: \code{\link{get_pump_test_vars}}}
}
\value{
combined pumptest and Q monitoring dataset
}
\description{
Combined Pumptest and Q Monitoring Dataset
}
