% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{correlation_plot}
\alias{correlation_plot}
\title{plots Qs_rel vs. input variable as box plot (categorical input variable)
or scatterplot (numerical input variable)}
\usage{
correlation_plot(df, x, y = "Qs_rel", title = gsub("_", " ", x))
}
\arguments{
\item{df}{data frame}

\item{x}{column name of x variable"}

\item{y}{column name of y variable (default Qs_rel")}
}
\description{
plots Qs_rel vs. input variable as box plot (categorical input variable)
or scatterplot (numerical input variable)
}
