% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{fill_up_na_with_median}
\alias{fill_up_na_with_median}
\title{Fill up NA values with median of lookup table}
\usage{
fill_up_na_with_median(df, df_lookup, matching_id = "well_id")
}
\arguments{
\item{df}{data frame with NA values}

\item{df_lookup}{data frame to calculate median values}

\item{matching_id}{column with ids for which median should be calculated}
}
\description{
Fill up NA values with median of lookup table
}
