% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{frequency_table}
\alias{frequency_table}
\title{calculate absolute and relative frequencies of categorical varables}
\usage{
frequency_table(x, perc_digits = 1)
}
\arguments{
\item{x}{vector with categorical variable}

\item{perc_digits}{number of decimal digits for percentages, default = 1}
}
\description{
calculate absolute and relative frequencies of categorical varables
}
