% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get Predictions}
\usage{
get_predictions(model, sim_data, predicted_parameter = "Qs_rel")
}
\arguments{
\item{model}{trained ML model}

\item{sim_data}{as retrieved by \code{\link{prepare_simulation_donothing_df}}}

\item{predicted_parameter}{name of predicted parameter (default: "Qs_rel")}
}
\value{
data frame with predictions
}
\description{
Get Predictions
}
