% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pump_test_vars.R
\name{get_pump_test_vars}
\alias{get_pump_test_vars}
\title{Get Default Pump Test Variables}
\usage{
get_pump_test_vars()
}
\value{
vector with column names of pump test variables
}
\description{
Get Default Pump Test Variables
}
\examples{
get_pump_test_vars()
}
