% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_Q_monitoring_data.R
\name{get_W_static_data}
\alias{get_W_static_data}
\title{Get W_static measurement data from Neubaupumpversuche, Kurzpumpversuche and other sources}
\usage{
get_W_static_data(path, renamings, df_wells)
}
\arguments{
\item{path}{path to static water level data (csv-file)}

\item{renamings}{list with renamings}

\item{df_wells}{data frame with prepared well data}
}
\description{
Get W_static measurement data from Neubaupumpversuche, Kurzpumpversuche and other sources
}
