% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{interpolate_and_fill}
\alias{interpolate_and_fill}
\title{Interpolate and fill up static water level}
\usage{
interpolate_and_fill(df, x_col, y_col, group_by_col, origin_col)
}
\arguments{
\item{df}{data frame}

\item{x_col}{x column, e.g. date, to be used for interpolation}

\item{y_col}{y column, e.g. measured values, to be used for interpolation}

\item{group_by_col}{grouping variable within which interpolation is done}

\item{origin_col}{already existing or to be created column with type of value}
}
\description{
Interpolate and fill up static water level
}
