% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{interpolate_Qs}
\alias{interpolate_Qs}
\title{Title}
\usage{
interpolate_Qs(df, interval_days = 1)
}
\arguments{
\item{df}{data frame with date and Qs measurements}

\item{interval_days}{interval for interpolation}
}
\description{
Title
}
