% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{load_renamings_csv}
\alias{load_renamings_csv}
\title{load renaming table from original excel file}
\usage{
load_renamings_csv(infile, fileEncoding = "WINDOWS-1252")
}
\arguments{
\item{infile}{full path to excel file}
}
\description{
load renaming table from original excel file
}
