% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_renamings_excel}
\alias{load_renamings_excel}
\title{load renaming table from original excel file}
\usage{
load_renamings_excel(
  infile,
  sheet = "DATEN",
  old_name_col = "Feld",
  new_name_col = "Parametername-R"
)
}
\arguments{
\item{infile}{full path to excel file}

\item{sheet}{sheet name}

\item{old_name_col}{name of column with original variable names}

\item{new_name_col}{name of column with new variable names}
}
\description{
load renaming table from original excel file
}
