% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu_model_data.R
\docType{data}
\name{model_data}
\alias{model_data}
\title{Data for well capacity prediction}
\format{
A data.frame with 6308 rows and 47 variables:
\describe{
\item{well_id}{well id, for info}
\item{date}{date of capacity measurement, for info}
\item{key}{measurement key, e.g. operational_start, pump_test_1, pump_test_2, for info}
\item{Qs_rel}{specific capacity of well relative to operational start condition, output}
\item{days_since_operational_start}{days since operational start, redundant}
\item{well_age_years}{years since operationa start, input, numeric}
\item{to be continued}{...}
}
}
\usage{
model_data
}
\description{
A dataset for well capacity prediction created with R script in /data-raw/model_data.R
}
\keyword{datasets}
