% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu_operational_start.R
\docType{data}
\name{operational_start}
\alias{operational_start}
\title{Input Data for Well Capacity Prediction: Operational Start}
\format{
A data.frame with 6308 rows and 2 variables:
\describe{
\item{well_id}{well id, for info}
\item{operational_start.date}{date of capacity measurement, for info}
}
}
\usage{
operational_start
}
\description{
Operational start date for all wells (prepared with R script in
/data-raw/prepare_model_data.R)
}
\examples{
head(operational_start, n = 20)
}
\keyword{datasets}
