% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_distribution}
\alias{plot_distribution}
\title{plot frequency distribution of numerical variable}
\usage{
plot_distribution(
  Data,
  variable,
  binwidth = NULL,
  title,
  vertical_x_axis_labels = TRUE,
  boundary = 0
)
}
\arguments{
\item{Data}{Data to be plotted}

\item{variable}{variable}

\item{binwidth}{binwidrh}

\item{title}{plot title}

\item{vertical_x_axis_labels}{should x-axis labels be ploted vertically (TRUE / FALSE)}

\item{boundary}{left boundary of bars, default: 0}
}
\description{
plot frequency distribution of numerical variable
}
