% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{plot frequency distribution of factor variable}
\usage{
plot_frequencies(
  Data,
  variable,
  title = variable,
  offset_perc_labels = 0.1,
  vertical_x_axis_labels = TRUE
)
}
\arguments{
\item{Data}{Data to be plotted}

\item{variable}{variable}

\item{title}{plot title}

\item{offset_perc_labels}{distance of labels from bars}

\item{vertical_x_axis_labels}{should x-axis labels be ploted vertically (TRUE / FALSE)}
}
\description{
plot frequency distribution of factor variable
}
