% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions_points_per_well.R
\name{plot_predictions_points_per_well}
\alias{plot_predictions_points_per_well}
\title{Plot Predictions: Points per Well}
\usage{
plot_predictions_points_per_well(
  predictions,
  max_simulation_years = NULL,
  dbg = TRUE
)
}
\arguments{
\item{predictions}{predictions as retrieved by \code{\link{get_predictions}}}

\item{max_simulation_years}{maximum years to be used as xlim, if not provided
(NULL) the full simulation period is used (default: NULL)}

\item{dbg}{print debug messages}
}
\value{
plots one page per well
}
\description{
Plot Predictions: Points per Well
}
