% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_Qs_over_time}
\alias{plot_Qs_over_time}
\title{plot_Qs_over_time}
\usage{
plot_Qs_over_time(df, xmax = 40, legend_position = "top")
}
\arguments{
\item{df}{data frame}

\item{xmax}{limit for maximum of x-axis (default: 40)}

\item{legend_position}{default: "top"}
}
\value{
plot Qs over time
}
\description{
plot_Qs_over_time
}
