% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pump_test_data.R
\name{prepare_pump_test_data_1}
\alias{prepare_pump_test_data_1}
\title{Prepare pump test data in wide format
Steps: i) read, rename and clean data, ii) correct wrong pump test dates,
iii) fill up missing pump test dates, iv) get information for replaced wells,
v) calculate Qs and Qs_rel, vi) determine action type, vii) select columns}
\usage{
prepare_pump_test_data_1(path, renamings, df_wells)
}
\arguments{
\item{path}{path to pump test data}

\item{renamings}{list with renamings}

\item{df_wells}{prepared data frame with well characteristics}
}
\description{
Prepare pump test data in wide format
Steps: i) read, rename and clean data, ii) correct wrong pump test dates,
iii) fill up missing pump test dates, iv) get information for replaced wells,
v) calculate Qs and Qs_rel, vi) determine action type, vii) select columns
}
