% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_quality_data.R
\name{prepare_quality_data}
\alias{prepare_quality_data}
\title{Prepare Quality Data}
\usage{
prepare_quality_data(path, renamings)
}
\arguments{
\item{path}{path}

\item{renamings}{renamings}
}
\value{
prepared quality day
}
\description{
Prepare Quality Data
}
