% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_simulation_donothing_df.R
\name{prepare_simulation_donothing_df}
\alias{prepare_simulation_donothing_df}
\title{Prepare 'Do Nothing' Simulation Data Frame}
\usage{
prepare_simulation_donothing_df(
  model_data = dwc.wells::model_data_reduced,
  rehabs = dwc.wells::rehabs,
  operational_start = dwc.wells::operational_start,
  sim_interval = 1L,
  sim_period = 60L,
  sim_reference_date = as.Date("2021-05-01"),
  predicted_parameter = "Qs_rel"
)
}
\arguments{
\item{model_data}{(reduced) model data set as retrieved by \code{\link{model_data_reduced}},
will be merged with but filtering out duplicated columns (i.e. 'well_age_years',
'n_rehab' and 'time_since_rehab_years') and name of "predicted_parameter"}

\item{rehabs}{rehabiliations dataset as retrieved by \code{\link{rehabs}}}

\item{operational_start}{operational start dataset as retrieved by
\code{\link{operational_start}}}

\item{sim_interval}{simulation interval in years at which predictions should be
performed (default: 1L)}

\item{sim_period}{simulation period in years since operational start of each
production well (default: 60L)}

\item{sim_reference_date}{(default: as.Date("2021-05-01")), used to generate
fill up column 'type' with 'past' (< sim_reference_date) and 'future'
(>= sim_reference_date)}

\item{predicted_parameter}{name of predicted parameter (default: "Qs_rel")}
}
\value{
simulation data frame
}
\description{
Prepares a simulation data frame with the regenerations that
were performed in the past but without assumung new ones in the future
}
