% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_volume_data.R
\name{prepare_volume_data}
\alias{prepare_volume_data}
\title{Prepare Volume Data}
\usage{
prepare_volume_data(path, renamings, df_wells)
}
\arguments{
\item{path}{path}

\item{renamings}{renamings}

\item{df_wells}{df_wells}
}
\value{
Prepared volume data
}
\description{
Prepare Volume Data
}
