% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{Qs_heatmap_plot}
\alias{Qs_heatmap_plot}
\title{Heatmap / raster plot for Qs values over time with each well as one line}
\usage{
Qs_heatmap_plot(
  df,
  colours,
  dummy_labels,
  date_limits,
  title,
  n_wells_per_page
)
}
\arguments{
\item{df}{data frame with date, well_id, Qs_rel}

\item{colours}{3 colours for low, middle and high colour limits}

\item{dummy_labels}{dummy labels if there are less wells than expected}

\item{date_limits}{vector with two date strings in format "yyyy-mm-dd"}

\item{title}{plot title}

\item{n_wells_per_page}{number of wells do be shown}
}
\description{
Heatmap / raster plot for Qs values over time with each well as one line
}
