% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{read_csv}
\alias{read_csv}
\title{Title read csv data file exported by Sebastian Schimmelpfennig from db2}
\usage{
read_csv(
  file,
  header = TRUE,
  fileEncoding = "UTF-8",
  skip = 2,
  dec = ".",
  sep = "\\t",
  na.strings = "(null)"
)
}
\arguments{
\item{file}{path to csv file}

\item{header}{logical, default = TRUE}

\item{fileEncoding}{default = UTF-8}

\item{skip}{number of rows to skip, default = 2}

\item{dec}{decimal separator, default = '.'}

\item{sep}{columns separator, default = 'tab'}

\item{na.strings}{string that represents NA, default = "(null)"}
}
\description{
Title read csv data file exported by Sebastian Schimmelpfennig from db2
}
