% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{read_ms_access}
\alias{read_ms_access}
\title{read table from MS Access data base via odbc connection under 64-bit-R}
\usage{
read_ms_access(path_db, tbl_name)
}
\arguments{
\item{path_db}{full path to database}

\item{tbl_name}{name of database table to be read}
}
\description{
read table from MS Access data base via odbc connection under 64-bit-R
}
