% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_select_rename}
\alias{read_select_rename}
\title{read table from MS Access data base; select and rename columns as defined in
renamings table ('old_name' -> 'new_name')}
\usage{
read_select_rename(path_db, tbl_name, renamings)
}
\arguments{
\item{path_db}{full path to database}

\item{tbl_name}{name of database table to be read}

\item{renamings}{name of data frame with renamings}
}
\description{
read table from MS Access data base; select and rename columns as defined in
renamings table ('old_name' -> 'new_name')
}
