% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu_rehabs.R
\docType{data}
\name{rehabs}
\alias{rehabs}
\title{Input Data for Well Capacity Prediction: Well Rehabilitations}
\format{
A data.frame with 6308 rows and 3 variables:
\describe{
\item{well_id}{well id, for info}
\item{n_rehab}{number of well rehabilitations}
\item{rehab_date}{date of well rehabilitation}
}
}
\usage{
rehabs
}
\description{
Well rehabilitations for all wells (prepared with R script in
/data-raw/prepare_model_data.R)
}
\keyword{datasets}
