% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{rename_values}
\alias{rename_values}
\title{rename values of a character vector according to renamings table}
\usage{
rename_values(
  x,
  renamings,
  old_name_col = "old_name",
  new_name_col = "new_name"
)
}
\arguments{
\item{x}{character vector}

\item{renamings}{data frame consisting of old and new names}

\item{old_name_col}{name of column with original variable names}

\item{new_name_col}{name of column with new variable names}
}
\description{
rename values of a character vector according to renamings table
}
