% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{replace_na_with_median}
\alias{replace_na_with_median}
\title{Replace NAs with median}
\usage{
replace_na_with_median(x)
}
\arguments{
\item{x}{vector, for which NA should be replaced}
}
\description{
Replace NAs with median
}
