% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_NAs_in_factor_var}
\alias{replace_NAs_in_factor_var}
\title{rename values of a character vector according to renamings table}
\usage{
replace_NAs_in_factor_var(x)
}
\arguments{
\item{x}{factor with NAs or gaps ("")}
}
\description{
rename values of a character vector according to renamings table
}
