% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_NAs_in_factor_vars}
\alias{replace_NAs_in_factor_vars}
\title{rename values of a character vector according to renamings table}
\usage{
replace_NAs_in_factor_vars(df)
}
\arguments{
\item{df}{data frame, with factors to be checked for NAs}
}
\description{
rename values of a character vector according to renamings table
}
