% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dataset.R
\name{resample_dataset}
\alias{resample_dataset}
\title{Resample Dataset}
\usage{
resample_dataset(model_data, method = "random", prop = 0.8)
}
\arguments{
\item{model_data}{(reduced) model data set as retrieved by \code{\link{model_data_reduced}}}

\item{method}{resampling method. Either "random" or "by_well" (default: "random")}

\item{prop}{The proportion of data to be retained for modeling/analysis (default:
0.8), i.e. 80\% of the dataset}
}
\value{
list with elements "training" and "testing" datasets
}
\description{
Resample Dataset
}
