% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{save_data}
\alias{save_data}
\title{Save data frame in different formats: csv, RData, rds}
\usage{
save_data(Data, path, filename)
}
\arguments{
\item{Data}{data frame}

\item{path}{out path for saving data}

\item{filename}{core of file name}
}
\description{
Save data frame in different formats: csv, RData, rds
}
