% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{scatterplot}
\alias{scatterplot}
\title{scatterplot for comparing numeric predictions with observations}
\usage{
scatterplot(df_pred, lines_80perc = FALSE)
}
\arguments{
\item{df_pred}{data frame obtained with tidymodels::collect_predictions() with
columns Qs_rel and .pred}

\item{lines_80perc}{logical value; shout 80\%-lines be drawn?; default = FALSE}
}
\description{
scatterplot for comparing numeric predictions with observations
}
