% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{select_rename_cols}
\alias{select_rename_cols}
\title{selects and renames columns from a data frame according to a reference table}
\usage{
select_rename_cols(
  df,
  renamings,
  old_name_col = "old_name",
  new_name_col = "new_name"
)
}
\arguments{
\item{df}{data frame with cols to be renamed}

\item{renamings}{name of data frame with renamings}

\item{old_name_col}{name of column with original variable names}

\item{new_name_col}{name of column with new variable names}
}
\description{
selects and renames columns from a data frame according to a reference table
}
