% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarise_marginal_factor_levels}
\alias{summarise_marginal_factor_levels}
\title{summarise factor levels with relative frequency below a threshold}
\usage{
summarise_marginal_factor_levels(x, perc_threshold, marginal_name)
}
\arguments{
\item{x}{factor variable}

\item{perc_threshold}{percentage threshold under which levels will be summarised}

\item{marginal_name}{for new summary factor level}
}
\description{
summarise factor levels with relative frequency below a threshold
}
