% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{load_renamings_csv}
\alias{load_renamings_csv}
\title{Load renaming table from CSV file}
\usage{
load_renamings_csv(infile, fileEncoding = "WINDOWS-1252", ...)
}
\arguments{
\item{infile}{full path to CSV file}

\item{fileEncoding}{Encoding in which the CSV file is given. Passed to
\code{\link{read.table}}. Defaut: "WINDOWS-1252"}

\item{\dots}{further arguments passed to \code{\link{read.table}}}
}
\description{
Load renaming table from CSV file
}
