% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rename_values}
\alias{rename_values}
\title{rename values of a character vector according to renamings table}
\usage{
rename_values(x, renamings)
}
\arguments{
\item{x}{character vector}

\item{renamings}{data frame consisting of columns old_name' and 'new_name'}
}
\description{
rename values of a character vector according to renamings table
}
