% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{save_data}
\alias{save_data}
\title{Save data frame in different formats: csv, RData, rds}
\usage{
save_data(Data, path, filename, formats = c("csv", "RData", "rds"))
}
\arguments{
\item{Data}{data frame}

\item{path}{out path for saving data}

\item{filename}{core of file name}

\item{formats}{export formats: "csv", "RData", "rds" or several using 'c'}
}
\description{
Save data frame in different formats: csv, RData, rds
}
